Move import functions to SESSA class
This prevents duplicate code and seems logical, as SESSA should always load everything provided by the configuration file
