+++ src/main/java/org/aksw/sessa/main/SESSA.java
@@ -1,20 +1,30 @@
 package org.aksw.sessa.main;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
+import java.util.stream.Stream;
 import org.aksw.sessa.candidate.Candidate;
 import org.aksw.sessa.candidate.CandidateGenerator;
 import org.aksw.sessa.colorspreading.ColorSpreader;
 import org.aksw.sessa.helper.files.handler.FileHandlerInterface;
+import org.aksw.sessa.helper.files.handler.RdfFileHandler;
+import org.aksw.sessa.helper.files.handler.ReverseTsvFileHandler;
+import org.aksw.sessa.helper.files.handler.TsvFileHandler;
 import org.aksw.sessa.helper.graph.GraphInterface;
 import org.aksw.sessa.helper.graph.Node;
 import org.aksw.sessa.importing.config.ConfigurationInitializer;
 import org.aksw.sessa.importing.config.exception.MalformedConfigurationException;
-import org.aksw.sessa.importing.dictionary.DictionaryInterface;
 import org.aksw.sessa.importing.dictionary.FileBasedDictionary;
 import org.aksw.sessa.importing.dictionary.energy.EnergyFunctionInterface;
+import org.aksw.sessa.importing.dictionary.energy.LevenshteinDistanceFunction;
+import org.aksw.sessa.importing.dictionary.energy.PageRankFunction;
 import org.aksw.sessa.importing.dictionary.implementation.HashMapDictionary;
 import org.aksw.sessa.importing.dictionary.implementation.LuceneDictionary;
 import org.aksw.sessa.importing.dictionary.util.Filter;
@@ -24,7 +34,9 @@ import org.aksw.sessa.query.models.QAModel;
 import org.aksw.sessa.query.processing.post.PostProcessing;
 import org.aksw.sessa.query.processing.pre.QueryProcessingInterface;
 import org.aksw.sessa.query.processing.pre.implementation.SimpleQueryProcessing;
+import org.apache.commons.configuration2.BaseHierarchicalConfiguration;
 import org.apache.commons.configuration2.Configuration;
+import org.apache.commons.configuration2.HierarchicalConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -36,6 +48,11 @@ public class SESSA {
   private static final Logger log = LoggerFactory.getLogger(SESSA.class);
 
   private static final String DICTIONARY_TYPE = "dictionary.type";
+  private static final String ENERGY_FUNCTION_KEY = "dictionary.energy_function";
+  private static final String FILES_KEY = "dictionary.files.location";
+  private static final String FILTER_LIMITS_KEY = "dictionary.filter.limits";
+  private static final String FILTER_NAMES_KEY = "dictionary.filter.names";
+  private static final String LUCENE_LOCATION_KEY = "dictionary.lucene.location";
   private static final String LUCENE_OVERRIDE_KEY = "dictionary.lucene.override_on_start";
 
   private FileBasedDictionary dictionary;
@@ -53,15 +70,31 @@ public class SESSA {
 
 
   /**
-   * Loads SESSA with the given configuration.
+   * Loads SESSA with the given configuration. All dictionaries in the configuration will also be
+   * imported.
    *
    * @param configuration configuration that should be applied to SESSA
    * @throws MalformedConfigurationException is thrown when the configuration file is malformed
    */
-  public SESSA(Configuration configuration) throws MalformedConfigurationException {
+  public SESSA(BaseHierarchicalConfiguration configuration) throws MalformedConfigurationException {
+    long startTime = System.nanoTime();
     queryProcess = new SimpleQueryProcessing();
     this.configuration = configuration;
     dictionary = initDictionary();
+    File f = new File(configuration.getString(LUCENE_LOCATION_KEY));
+    if (configuration.getBoolean(LUCENE_OVERRIDE_KEY) &&
+        configuration.getString(DICTIONARY_TYPE).equals("lucene") &&
+        f.exists()) {
+      log.info("Skipping building dictionary.");
+    } else {
+      log.info("Building dictionary from files. This could take some time!");
+      loadDictionaries(configuration);
+      long endTime = System.nanoTime();
+      log.info("Finished importing dictionary (in {}sec).",
+          (endTime - startTime) / (1000 * 1000 * 1000));
+    }
+    addFilters(configuration);
+    applyEnergyFunction(configuration);
   }
 
   /**
@@ -207,5 +240,99 @@ public class SESSA {
         throw new MalformedConfigurationException(
             String.format("Could not determine value of property '%s'", DICTIONARY_TYPE));
     }
+  }
+
+
+  private void loadDictionaries(BaseHierarchicalConfiguration configuration) {
+    HierarchicalConfiguration subConfig = configuration.configurationAt(FILES_KEY);
+    if (subConfig.containsKey("rdf")) {
+      log.info("Found entry for rdf-files in configuration file. Importing...");
+      loadSingleDictionary(new RdfFileHandler(), subConfig.getString("rdf"));
+    }
+    if (subConfig.containsKey("tsv")) {
+      log.info("Found entry for tsv-files in configuration file. Importing...");
+      loadSingleDictionary(new TsvFileHandler(), subConfig.getString("tsv"));
+    }
+    if (subConfig.containsKey("reverse_tsv")) {
+      log.info("Found entry for reverse tsv-files in configuration file. Importing...");
+      loadSingleDictionary(new ReverseTsvFileHandler(), subConfig.getString("reverse_tsv"));
+    }
+  }
+
+
+  private void loadSingleDictionary(FileHandlerInterface handler,
+      String pathString) {
+    log.debug("Path to files is '{}'", pathString);
+    try (Stream<Path> path = Files.walk(Paths.get(pathString))) {
+      path
+          .filter(Files::isRegularFile)
+          .forEach((Path file) -> {
+            try {
+              log.info("Loading file '{}' to dictionary via {}.",
+                  file.toString(),
+                  handler.getClass().getSimpleName());
+              handler.loadFile(file.toString());
+              loadFileToDictionary(handler);
+            } catch (IOException ioE) {
+              log.error(ioE.getLocalizedMessage());
+            }
+          });
+    } catch (IOException ioE) {
+      log.warn(
+          "Could not load any file in given path '{}'.", pathString);
+    }
+  }
+
+  /**
+   * Use this method for adding filters and manipulate the energy score.
+   */
+  private void addFilters(BaseHierarchicalConfiguration configuration)
+      throws MalformedConfigurationException {
+    String[][] filters = new String[2][];
+    filters[0] = configuration.getStringArray(FILTER_NAMES_KEY);
+    filters[1] = configuration.getStringArray(FILTER_LIMITS_KEY);
+    if (filters[0].length < filters[1].length) {
+      log.info(
+          "The number of filters and limits in the configuration is not the same! Using smaller value.");
+    }
+    int length = filters[0].length < filters[1].length ? filters[0].length : filters[1].length;
+    for (int i = 0; i < length; i++) {
+      EnergyFunctionInterface function = getFunction(filters[0][i]);
+      int limit;
+      try {
+        limit = Integer.parseInt(filters[1][i]);
+        log.debug("\t Limit is {}", limit);
+      } catch (NumberFormatException nfE) {
+        throw new MalformedConfigurationException(String
+            .format("Error in %s: Given limit is not an integer. #%d:%s", FILTER_LIMITS_KEY, i,
+                filters[1][i]));
+      }
+      Filter filter = new Filter(function, limit);
+      addFilter(filter);
+    }
+  }
+
+  private EnergyFunctionInterface getFunction(String functionName)
+      throws MalformedConfigurationException {
+    switch (functionName) {
+      case "levenshtein":
+        log.debug("Add Levenshtein filter.");
+        return new LevenshteinDistanceFunction();
+      case "pagerank":
+        log.debug("Add PageRank filter.");
+        return new PageRankFunction();
+      default:
+        throw new MalformedConfigurationException(
+            String.format("Could not determine value of property '%s'. Given value: %s",
+                FILTER_NAMES_KEY, functionName));
+    }
+  }
+
+
+  private void applyEnergyFunction(BaseHierarchicalConfiguration configuration)
+      throws MalformedConfigurationException {
+    String energyFunctionName = configuration.getString(ENERGY_FUNCTION_KEY);
+    EnergyFunctionInterface lFunction = getFunction(energyFunctionName);
+    setEnergyFunction(lFunction);
   }
 }
